week07/src/binaryTree/BinaryTreeTest.java
package binaryTree;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class BinaryTreeTest {

	@BeforeAll
	static void setUpBeforeClass() throws Exception {
		
	}

	@AfterAll
	static void tearDownAfterClass() throws Exception {
		
	}

	@BeforeEach
	void setUp() throws Exception {
		
	}

	@AfterEach
	void tearDown() throws Exception {
		
	}

	@Test
	void test() {
		BinaryTree<Integer> tree = new BinaryTree<>();
		tree.insert(1);
		tree.insert(0);
		tree.insert(2);

		Integer topValue = 1;
		Integer leftValue = 0;
		Integer rightValue = 2;
		assertEquals(topValue, tree.value());
		assertEquals(leftValue, tree.left().value());
		assertEquals(rightValue, tree.right().value());
	}
}
